refactor(api): move security pre-handler mocks
