+++ api/tests/devcomp/integration/application/trainings/training-route_test.js
@@ -3,7 +3,7 @@ import { trainingController } from '../../../../../src/devcomp/application/train
 import * as moduleUnderTest from '../../../../../src/devcomp/application/trainings/training-route.js';
 import { securityPreHandlers } from '../../../../../src/shared/application/security-pre-handlers.js';
 import { expect, HttpTestServer, sinon } from '../../../../test-helper.js';
-import { getAdminRoleStub } from '../../../../tooling/security-pre-handlers-helpers.js';
+import { getAdminRoleStub } from '../../../../tooling/mocks/security-pre-handlers.mock.js';
 
 describe('Integration | Devcomp | Application | Trainings | Router | training-router', function () {
   describe('GET /api/admin/trainings/${trainingId}', function () {
